Enhance database initialization and shell script functionality

+++ scripts/enter-shell.sh
@@ -6,4 +6,10 @@ source ./scripts/common.sh
 validate_service $1
 load_env
 clear_screen
-docker compose -f compose.$ENV.yaml --progress quiet run --rm $1 bash
+if [ "$1" = "db" ] || [ "$1" = "redis" ]; then
+    # For db and redis, creating a new container will make it unable to connect to postgres/redis
+    # So we need to use exec to get a shell from the existing container.
+    docker compose -f compose.$ENV.yaml --progress quiet exec $1 bash
+else
+    docker compose -f compose.$ENV.yaml --progress quiet run --rm $1 bash
+fi
